+++ apps/ws/src/stream/sendStreamState.ts
@@ -1,0 +1,6 @@
+
+export function sendStreamState(parsedMessege:any){
+    // Send the state of the redis to the user 
+    console.log("sent");
+
+}
